requirements.txt
celery
Django==2.0.6
djangorestframework==3.8.2
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
